requirements.txt
django==5.2
djangorestframework==3.16.0
Pillow==11.2.1
gunicorn==23.0.0